Align LocateAnything tests with model suite
